requirements.txt
numpy==1.26.1
Pillow==10.1.0
requests==2.31.0
scipy==1.11.3
PyAudio~=0.2.11